Update MainPage css, Refactor
